requirements.txt
-f https://download.pytorch.org/whl/cu130
torch==2.10.0
torchvision==0.25.0
torchaudio==2.10.0

torchdata
psutil

scanpy==1.11.5
datasets
pytorch-lightning==2.2.1
tensorboard==2.14.0
einops
omegaconf

scib
scipy
pandas
numpy


ninja

adjustText
matplotlib_venn
pexpect
setuptools<82

